:art: Split panel element into a separate file
